# docs/swagger.yaml
basePath: /
definitions:
  domain.Branch:
    properties:
      Description:
        example: description
        type: string
      address:
        example: Pontianak
        type: string
      code:
        example: PTK
        type: string
      created_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      id:
        example: 953c294b-4535-4656-9fd8-be58zzd0402a9a1x
        type: string
      name:
        example: Pontianak
        type: string
    type: object
  domain.CharityMrys:
    properties:
      Amount:
        example: 40000
        type: integer
      Description:
        example: description
        type: string
      Month:
        example: 1
        type: integer
      Userame:
        example: usename
        type: string
      Year:
        example: 2021
        type: integer
      branch:
        $ref: '#/definitions/domain.Branch'
      created_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      id:
        example: "1"
        type: string
      name:
        example: Leo Messi
        type: string
      user_id:
        example: "1"
        type: string
    type: object
  domain.CharityMrysAll:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.CharityMrys'
        type: array
    type: object
  domain.CharityMrysPagination:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.CharityMrys'
        type: array
      meta:
        $ref: '#/definitions/domain.MetaPagination'
    type: object
  domain.MetaPagination:
    properties:
      currentPage:
        type: integer
      offset:
        type: integer
      perPage:
        type: integer
      sort:
        type: string
      total:
        type: integer
      totalPage:
        type: integer
    type: object
  domain.ReceiptLunar:
    properties:
      branch:
        $ref: '#/definitions/domain.Branch'
      created_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      description:
        type: string
      id:
        example: "1"
        type: string
      internation_date:
        example: "2021-08-04"
        type: string
      lunar_date:
        example: "2021-08-04"
        type: string
      receipt_lunar_detail:
        items:
          $ref: '#/definitions/domain.ReceiptLunarDetail'
        type: array
      total:
        example: 40000
        type: integer
      user_id:
        example: "1"
        type: string
      username:
        example: username
        type: string
    type: object
  domain.ReceiptLunarDetail:
    properties:
      created_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      description:
        type: string
      id:
        example: "1"
        type: string
      name:
        type: string
      receipt_lunar:
        $ref: '#/definitions/domain.ReceiptLunar'
      total:
        example: 40000
        type: integer
    type: object
  domain.ReceiptLunarPagination:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.ReceiptLunar'
        type: array
      meta:
        $ref: '#/definitions/domain.MetaPagination'
    type: object
  usecase.CreateBulkCharityMrysInput:
    properties:
      amount:
        example: 40000
        type: integer
      branch:
        $ref: '#/definitions/domain.Branch'
      description:
        example: description
        type: string
      id:
        example: "1"
        type: string
      month_from:
        example: 2
        type: integer
      month_to:
        example: 10
        type: integer
      name:
        example: Leo Messi
        type: string
      year:
        example: 2021
        type: integer
    required:
    - amount
    - month_from
    - month_to
    - name
    - year
    type: object
  usecase.CreateReceiptLunarInput:
    properties:
      branch:
        $ref: '#/definitions/domain.Branch'
      description:
        type: string
      id:
        type: string
      international_date:
        type: string
      lunar_date:
        type: string
      receipt_lunar_detail:
        items:
          $ref: '#/definitions/domain.ReceiptLunarDetail'
        type: array
    required:
    - international_date
    - lunar_date
    type: object
  usecase.DeleteOneCharityMrysOutput:
    properties:
      id:
        type: string
      success:
        type: boolean
    type: object
  usecase.DeleteOneReceiptLunarOutput:
    properties:
      id:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8000/vhry/data/
info:
  contact:
    email: felixanthony1996.fa@gmail.com
  description: Data API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Data API
  version: "1.0"
paths:
  /v1/charity-mrys:
    post:
      consumes:
      - application/json
      description: Create Charity Mrys
      parameters:
      - description: Create charity mrys
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/domain.CharityMrys'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.CharityMrys'
      security:
      - ApiKeyAuth: []
      summary: Create Charity Mrys
      tags:
      - CharityMrys
  /v1/charity-mrys/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Charity Mrys ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.DeleteOneCharityMrysOutput'
      security:
      - ApiKeyAuth: []
      summary: Delete One Charity Mrys By ID
      tags:
      - CharityMrys
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CharityMrys'
      security:
      - ApiKeyAuth: []
      summary: Find One Charity Mrys By ID
      tags:
      - CharityMrys
    patch:
      consumes:
      - application/json
      description: Update Charity Mrys By ID
      parameters:
      - description: Update charity mrys
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/domain.CharityMrys'
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.CharityMrys'
      security:
      - ApiKeyAuth: []
      summary: Update Charity Mrys By ID
      tags:
      - CharityMrys
  /v1/charity-mrys/create-bulk:
    post:
      consumes:
      - application/json
      description: Create Bulk Charity Mrys
      parameters:
      - description: Create charity mrys
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/usecase.CreateBulkCharityMrysInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/domain.CharityMrys'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Create Bulk Charity Mrys
      tags:
      - CharityMrys
  /v1/charity-mrys/list-all:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CharityMrysAll'
      security:
      - ApiKeyAuth: []
      summary: Find All CharityMrys
      tags:
      - CharityMrys
  /v1/charity-mrys/list-pagination/{currentPage}/{perPage}/{sort}:
    get:
      consumes:
      - application/json
      parameters:
      - description: CurrentPage
        in: path
        name: currentPage
        required: true
        type: integer
      - description: PerPage
        in: path
        name: perPage
        required: true
        type: integer
      - description: Sort
        in: path
        name: sort
        required: true
        type: string
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CharityMrysPagination'
      security:
      - ApiKeyAuth: []
      summary: Find Pagination CharityMrys
      tags:
      - CharityMrys
  /v1/receipt-lunar:
    post:
      consumes:
      - application/json
      description: Create Receipt Lunar
      parameters:
      - description: Create receipt lunar
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/usecase.CreateReceiptLunarInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.ReceiptLunar'
      security:
      - ApiKeyAuth: []
      summary: Create Receipt Lunar
      tags:
      - ReceiptLunar
  /v1/receipt-lunar/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Receipt Lunar ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.DeleteOneReceiptLunarOutput'
      security:
      - ApiKeyAuth: []
      summary: Delete One Receipt Lunar By ID
      tags:
      - ReceiptLunar
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ReceiptLunar'
      security:
      - ApiKeyAuth: []
      summary: Find One ReceiptLunar By ID
      tags:
      - ReceiptLunar
    patch:
      consumes:
      - application/json
      description: Update ReceiptLunar By ID
      parameters:
      - description: Update Receipt Lunar
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/domain.ReceiptLunar'
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.ReceiptLunar'
      security:
      - ApiKeyAuth: []
      summary: Update Receipt Lunar By ID
      tags:
      - ReceiptLunar
  /v1/receipt-lunar/list-pagination/{currentPage}/{perPage}/{sort}:
    get:
      consumes:
      - application/json
      parameters:
      - description: CurrentPage
        in: path
        name: currentPage
        required: true
        type: integer
      - description: PerPage
        in: path
        name: perPage
        required: true
        type: integer
      - description: Sort
        in: path
        name: sort
        required: true
        type: string
      - description: Search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ReceiptLunarPagination'
      security:
      - ApiKeyAuth: []
      summary: Find Pagination ReceiptLunar
      tags:
      - ReceiptLunar
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"